update sale tests
